windows_use_autonomous_agent/pyproject.toml
[project]
name = "windows-use"
version = "0.1.31"
description = "An AI Agent that interacts with Windows OS at GUI level."
readme = "readme.md"
authors = [
    { name = "Jeomon George", email = "jeogeoalukka@gmail.com" }
]
license = 'MIT'
license-files = ["LICENSE"]
urls = { homepage = "https://github.com/CursorTouch" }
keywords = ["windows", "agent", "ai", "desktop","ai agent","automation"]
requires-python = ">=3.12"
dependencies = [
    # Core Dependencies
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "pydantic>=2.5.0",
    
    # Data Processing & ML
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    
    # Web & HTTP
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.1",
    
    # UI Automation & Desktop
    "uiautomation>=2.0.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "humancursor>=0.1.0",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    
    # Image Processing
    "Pillow>=10.0.0",
    
    # System & Process Management
    "psutil>=5.9.0",
    
    # Text Processing
    "markdownify>=0.11.6",
    
    # GUI & Visualization
    "matplotlib>=3.7.0",
    
    # Utilities
    "termcolor>=2.3.0",
    "setuptools>=68.0.0",
    "twine>=4.0.2",
]

[project.optional-dependencies]
voice = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "webrtcvad>=2.0.10",
]
office = [
    "pywin32>=306",
    "openpyxl>=3.1.0",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["windows_use"]
include-package-data = true

[tool.setuptools.package-data]
"windows_use.agent.prompts" = ["*.md"]
